Add function for pre-populating fact extraction terms into DB
